Token auth setup with master for imports

+++ app/main.py
@@ -1,7 +1,21 @@
-from fastapi import FastAPI, HTTPException, status
+from fastapi import Depends, FastAPI, HTTPException, Request, status
 from fastapi.middleware.cors import CORSMiddleware
-from mongoengine import Document, FloatField, IntField, StringField, connect, errors
+from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm
+from mongoengine import (
+    Document,
+    BooleanField,
+    DateTimeField,
+    FloatField,
+    IntField,
+    StringField,
+    connect,
+    errors,
+)
+
+from datetime import datetime
+from secrets import token_urlsafe
 from pydantic import BaseModel
+import os
 
 # Fast API main app
 app = FastAPI()
@@ -17,6 +31,7 @@ app.add_middleware(
 )
 
 # Connect to mongo
+# TODO: renameDB
 connect("trees", host="mongodb://bosmapper_mongo")
 
 
@@ -79,7 +94,7 @@ class TreeDB(Document):
     notes = StringField(max_length=300)
 
 
-class SpeciesJson(BaseModel):
+class ImportSpeciesJson(BaseModel):
     species: list
     updated: str
 
@@ -159,19 +174,33 @@ def trees_json():
 
 
 @app.get("/trees/clear/")
-def remove_all():
+def remove_all(request: Request):
     """
     Remove all trees
     """
+    # Check for master token
+    if not request.headers.get("Master") == os.environ.get("MASTER_TOKEN"):
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Invalid authentication token",
+        )
+
     TreeDB.objects.all().delete()
     return {"detail": "All trees removed from collection"}
 
 
 @app.post("/trees/import/")
-def import_geojson(geojson: GeoJson):
+def import_geojson(geojson: GeoJson, request: Request):
     """
     Import trees from GeoJSON
     """
+    # Check for master token
+    if not request.headers.get("Master") == os.environ.get("MASTER_TOKEN"):
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Invalid authentication token",
+        )
+
     TreeDB.objects.all().delete()
 
     for feature in geojson.features:
@@ -282,10 +311,18 @@ def species_json():
 
 
 @app.post("/species/import/")
-def import_species(species_json: SpeciesJson):
+def import_species(species_json: ImportSpeciesJson, request: Request):
     """
     Import trees from GeoJSON
     """
+
+    # Check for master token
+    if not request.headers.get("Master") == os.environ.get("MASTER_TOKEN"):
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Invalid authentication token",
+        )
+
     SpeciesDB.objects.all().delete()
 
     for item in species_json.species:
@@ -293,3 +330,98 @@ def import_species(species_json: SpeciesJson):
         new_species.save()
 
     return {"detail": f"Imported {len(species_json.species)} species"}
+
+
+# AUTHENTICATION
+oauth2_scheme = OAuth2PasswordBearer(tokenUrl="/token")
+
+
+class ImportUsersJson(BaseModel):
+    passcodes: list
+
+
+class User(BaseModel):
+    passcode: str
+    token: str
+    token_generated: datetime
+    disabled: bool = False
+
+
+class UsersDB(Document):
+    """
+    Mongo user schema
+    """
+
+    passcode = StringField(max_length=30)
+    token = StringField(max_length=30)
+    token_generated = DateTimeField()
+    disabled = BooleanField()
+
+
+@app.post("/users/import/")
+def import_users(users_json: ImportUsersJson, request: Request):
+    """
+    Import users from json
+    """
+
+    # Check for master token
+    if not request.headers.get("Master") == os.environ.get("MASTER_TOKEN"):
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Invalid authentication token",
+        )
+
+    UsersDB.objects.all().delete()
+    for passcode in users_json.passcodes:
+        new_user = UsersDB(
+            passcode=passcode, token=token_urlsafe(20), token_generated=datetime.now(),
+        )
+        new_user.save()
+
+    return {"detail": f"Imported {len(users_json.passcodes)} items from passcode list"}
+
+
+def get_user(token: str):
+    try:
+        db_user = UsersDB.objects.get(token=token)
+        return User(**db_user.to_mongo())
+
+    except UsersDB.DoesNotExist:
+        return None
+
+
+async def get_current_user(token: str = Depends(oauth2_scheme)):
+    user = get_user(token)
+    if not user:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Invalid authentication credentials",
+            headers={"WWW-Authenticate": "Bearer"},
+        )
+    return user
+
+
+async def get_current_active_user(current_user: User = Depends(get_current_user)):
+    if current_user.disabled:
+        raise HTTPException(status_code=400, detail="Inactive user")
+    return current_user
+
+
+@app.post("/token/")
+async def login(form_data: OAuth2PasswordRequestForm = Depends()):
+    try:
+        user = UsersDB.objects.get(passcode=form_data.password)
+
+    except UsersDB.DoesNotExist:
+        raise HTTPException(status_code=400, detail="Incorrect credentials")
+
+    return {
+        "access_token": user.token,
+        "token_type": "bearer",
+        "token_generated": user.token_generated,
+    }
+
+
+@app.get("/users/me/")
+async def read_users_me(current_user: User = Depends(get_current_active_user)):
+    return current_user
